HashedPointMap is only needed PointMapCollection
