+++ staging_vespalib/src/vespa/vespalib/metrics/point_map.h
@@ -10,18 +10,5 @@ namespace metrics {
 
 using PointMap = std::map<Dimension, Label>;
 
-// internal
-class HashedPointMap {
-private:
-    const PointMap _map;
-    size_t _hash;
-public:
-    HashedPointMap() : _map(), _hash(0) {}
-    HashedPointMap(PointMap &&from);
-    bool operator< (const HashedPointMap &other) const;
-
-    const PointMap &backingMap() const { return _map; }
-};
-
 } // namespace vespalib::metrics
 } // namespace vespalib
