+++ staging_vespalib/src/vespa/vespalib/metrics/point_map_collection.h
@@ -10,6 +10,19 @@ namespace vespalib {
 namespace metrics {
 
 // internal
+class HashedPointMap {
+private:
+    const PointMap _map;
+    size_t _hash;
+public:
+    HashedPointMap() : _map(), _hash(0) {}
+    HashedPointMap(PointMap &&from);
+    bool operator< (const HashedPointMap &other) const;
+
+    const PointMap &backingMap() const { return _map; }
+};
+
+// internal
 class PointMapCollection {
 private:
     using PointMapMap = std::map<HashedPointMap, size_t>;
@@ -18,8 +31,8 @@ private:
     PointMapMap _map;
     std::vector<PointMapMap::const_iterator> _vec;
 public:
-    const HashedPointMap &lookup(size_t id) const;
-    size_t resolve(HashedPointMap map);
+    const PointMap &lookup(size_t id) const;
+    size_t resolve(PointMap map);
     size_t size() const;
 
     PointMapCollection() = default;
